get_meta_description() moved from ContextualData to Sourced
